Reorganization and cleaning of entry point and loader functions
Try to keep function related to the entry points separate from
the plugin loader that simply leverages those entry point to
load plugin classes
