+++ aiida/backends/djsite/db/models.py
@@ -185,12 +185,12 @@ class DbNode(m.Model):
         appropriate subclass.
         """
         from aiida.orm.node import Node
-        from aiida.plugins.loader import type_string_to_plugin_type
+        from aiida.plugins.loader import get_plugin_type_from_type_string
         from aiida.common import aiidalogger
         from aiida.plugins.loader import load_plugin
 
         try:
-            pluginclassname = type_string_to_plugin_type(self.type)
+            pluginclassname = get_plugin_type_from_type_string(self.type)
         except DbContentError:
             raise DbContentError("The type name of node with pk= {} is "
                                  "not valid: '{}'".format(self.pk, self.type))
