+++ aiida/cmdline/commands/group.py
@@ -197,7 +197,7 @@ class Group(VerdiCommandWithSubcommands):
         from aiida.orm import Group as G
         from aiida.common.utils import str_timedelta
         from aiida.utils import timezone
-        from aiida.plugins.loader import type_string_to_plugin_type
+        from aiida.plugins.loader import get_plugin_type_from_type_string
         from tabulate import tabulate
 
         parser = argparse.ArgumentParser(
@@ -271,7 +271,7 @@ class Group(VerdiCommandWithSubcommands):
                 if parsed_args.uuid:
                     row.append(n.uuid)
                 row.append(n.pk)
-                row.append(type_string_to_plugin_type(n.dbnode.type).
+                row.append(get_plugin_type_from_type_string(n.dbnode.type).
                            rsplit(".", 1)[1])
 
                 row.append(str_timedelta(now - n.ctime, short=True,
