+++ aiida/orm/implementation/django/calculation/job/__init__.py
@@ -18,7 +18,7 @@ from aiida.backends.djsite.utils import get_automatic_user
 from aiida.orm.group import Group
 from aiida.orm.implementation.django.calculation import Calculation
 from aiida.orm.implementation.general.calculation.job import AbstractJobCalculation
-from aiida.plugins.loader import type_string_to_plugin_type
+from aiida.plugins.loader import get_plugin_type_from_type_string
 from aiida.utils import timezone
